# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Priya Global Platform API
  version: 1.0.0
  description: |
    # Priya Global - AI-Powered Multi-Channel Sales Platform

    Complete API documentation for the Priya Global Platform - a comprehensive multi-tenant system for managing customer communications, AI-powered chat, billing, analytics, and business operations across 36 microservices.

    ## Key Features
    - **Multi-tenant Architecture**: Fully isolated workspaces with granular access control
    - **36 Microservices**: Specialized services for auth, channels, AI, billing, analytics, and more
    - **Multi-channel Support**: WhatsApp, Email, Voice, SMS, Social, WebChat, Telegram, RCS, Video
    - **AI Integration**: Advanced conversational AI, sentiment analysis, and intelligent routing
    - **Enterprise Billing**: Stripe-powered subscriptions with regional pricing (USD, INR, GBP, AUD)
    - **Real-time Analytics**: Comprehensive dashboards and custom reporting
    - **Webhooks & Callbacks**: Event-driven architecture for real-time integrations
    - **Rate Limiting**: Plan-based throttling with per-tenant limits

    ## Authentication
    All endpoints use JWT (Bearer token) authentication except public webhooks.

    ```
    Authorization: Bearer {access_token}
    ```

    Access tokens are valid for 15 minutes. Use the refresh token to obtain new access tokens.

    ## Multi-tenancy
    Include tenant context in all requests:
...

    ## Rate Limiting
    Rates are plan-based:
...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/conversations:
    get:
      summary: List conversations
      tags: [Conversations]
      description: Retrieve paginated list of conversations with optional filtering
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, closed, archived]
        - name: channel
          in: query
          schema:
            type: string
            enum: [whatsapp, email, sms, voice, webchat, telegram, social]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at, updated_at, last_message_at]
      responses:
        '200':
          description: List of conversations
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Conversation'
                  total_count:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer

    post:
      summary: Create conversation
      tags: [Conversations]
      description: Initiate a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contact_id, channel]
              properties:
                contact_id:
                  type: string
                  format: uuid
                channel:
                  type: string
                  enum: [whatsapp, email, sms, voice, webchat, telegram, social]
                subject:
                  type: string
                initial_message:
                  type: string
                metadata:
                  type: object
      responses:
        '201':
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/channels:
    get:
      summary: List configured channels
      tags: [Channels]
      description: Get all active communication channels
      responses:
        '200':
          description: List of channels
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Channel'

    post:
      summary: Configure channel
      tags: [Channels]
      description: Add new communication channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, credentials]
              properties:
                type:
                  type: string
                  enum: [whatsapp, email, sms, voice, telegram, webchat, social, rcs, video]
                credentials:
                  type: object
                  description: Channel-specific credentials
                name:
                  type: string
                enabled:
                  type: boolean
                  default: true
      responses:
        '201':
          description: Channel configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '400':
          description: Invalid configuration
...
  /api/v1/analytics/conversations:
    get:
      summary: Conversation analytics
      tags: [Analytics]
      description: Detailed conversation metrics and trends
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: group_by
          in: query
          schema:
            type: string
            enum: [day, week, month, channel, agent]
      responses:
        '200':
          description: Conversation analytics
          content:
            application/json:
              schema:
                type: object
                properties:
                  metrics:
                    type: array
                    items:
                      type: object
...
  /api/v1/workflows:
    get:
      summary: List workflows
      tags: [Workflows]
      description: Get automation workflows
      responses:
        '200':
          description: List of workflows
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workflow'

    post:
      summary: Create workflow
      tags: [Workflows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, trigger, actions]
              properties:
                name:
                  type: string
                trigger:
                  type: object
                actions:
                  type: array
                  items:
                    type: object
      responses:
        '201':
          description: Workflow created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
...
  /api/v1/leads:
    get:
      summary: List leads
      tags: [Leads]
      description: Get captured leads
      responses:
        '200':
          description: List of leads
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Lead'

    post:
      summary: Capture lead
      tags: [Leads]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, contact]
              properties:
                name:
                  type: string
                contact:
                  type: object
                  required: [email]
                  properties:
                    email:
                      type: string
                    phone:
                      type: string
                source:
                  type: string
      responses:
        '201':
          description: Lead captured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token for authentication. Obtain via /api/v1/auth/login
  schemas:
    Error:
      type: object
      required: [detail, code]
      properties:
        detail:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
        request_id:
          type: string
          format: uuid
          description: Unique request identifier for tracking
        timestamp:
          type: string
          format: date-time
...
    Conversation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        contact_id:
          type: string
          format: uuid
        channel:
          type: string
        status:
          type: string
        subject:
          type: string
        assigned_to:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        message_count:
          type: integer
        last_message_at:
          type: string
          format: date-time
...
    Channel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        name:
          type: string
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    Workflow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        enabled:
          type: boolean
        trigger:
          type: object
        actions:
          type: array
        created_at:
          type: string
          format: date-time
...
    Lead:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        source:
          type: string
        created_at:
          type: string
          format: date-time